blockchainvibe/server/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication
PyJWT==2.8.0
python-multipart==0.0.6

# AI and News Processing
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3

# CORS (built into FastAPI)
# fastapi-cors==0.0.6  # Not needed, FastAPI has built-in CORS